refactor: move services to their own file, feature: implement team repository basics
